1、修改NodeInfo接口属性的类型(int -> Integer, root -> Long)； 2、添加返回Adiacency List(临接表)接口，返回id parent_id结构; 3、新增测试类;

+++ src/main/java/org/code_factory/jpa/nestedset/NodeInfo.java
@@ -17,12 +17,12 @@ package org.code_factory.jpa.nestedset;
  */
 public interface NodeInfo {
     Long getId();
-    int getLeftValue();
-    int getRightValue();
-    int getLevel();
-    int getRootValue();
-    void setLeftValue(int value);
-    void setRightValue(int value);
-    void setLevel(int level);
-    void setRootValue(int value);
+    Integer getLeftValue();
+    Integer getRightValue();
+    Integer getLevel();
+    Long getRootValue();
+    void setLeftValue(Integer value);
+    void setRightValue(Integer value);
+    void setLevel(Integer level);
+    void setRootValue(Long value);
 }
